Enabled Modular design for complex application
