+++ backend/app/__init__.py
@@ -1,0 +1,6 @@
+"""
+News Advisor AI Fact Checker
+A modular Flask application for fact-checking claims using web search and ML models.
+"""
+
+__version__ = "2.0.0"
